collector/requirements.txt
requests==2.31.0  
pandas==2.2.0  
psycopg2-binary==2.9.9  
sqlalchemy==2.0.27  
python-dotenv==1.0.1  
beautifulsoup4==4.12.2
lxml==4.9.3